requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
pydantic>=1.10.0,<2.0.0
uvicorn[standard]>=0.24.0

